Begin refactor, split wgridder from nifty directory
