+++ africanus/gridding/nifty/dask.py
@@ -28,514 +28,6 @@ else:
     nifty_import_err = None
 
 from africanus.util.requirements import requires_optional
-from ducc0.wgridder import ms2dirty, dirty2ms
-import dask
-from daskms import xds_from_ms, xds_from_table, xds_to_table, Dataset
-from numpy.testing import assert_array_equal
-from scipy.fft import next_fast_len
-
-
-def _dot_wrapper(uvw, freq, model, freq_bin_idx, freq_bin_counts,
-                 cellx, celly, nu, nv, epsilon, nthreads, do_wstacking, ncorr, complex_type):
-    return _dot_internal(uvw[0], freq, model[0][0],
-                         freq_bin_idx, freq_bin_counts,
-                         cellx, celly, nu, nv, epsilon, nthreads,
-                         do_wstacking, ncorr, complex_type)
-
-def _dot_internal(uvw, freq, model, freq_bin_idx, freq_bin_counts,
-                  cellx, celly, nu, nv, epsilon, nthreads, do_wstacking, ncorr, complex_type):
-    freq_bin_idx -= freq_bin_idx.min()  # adjust for chunking
-    nband = freq_bin_idx.size
-    nrow = uvw.shape[0]
-    nchan = freq.size
-    vis = np.zeros((nrow, nchan, ncorr), dtype=complex_type)
-    for i in range(nband):
-        I = slice(freq_bin_idx[i], freq_bin_idx[i] + freq_bin_counts[i])
-        vis[:, I, 0] = dirty2ms(uvw=uvw, freq=freq[I], dirty=model[i], wgt=None, 
-                                pixsize_x=cellx, pixsize_y=celly, nu=nu, nv=nv,
-                                epsilon=epsilon, nthreads=nthreads, 
-                                do_wstacking=do_wstacking)
-        vis[:, I, -1] = vis[:, I, 0]  # assume no Stokes Q for now
-    return vis
-
-def _hdot_wrapper(uvw, freq, ms, wgt, freq_bin_idx, freq_bin_counts,
-                  nx, ny, cellx, celly, nu, nv, epsilon, nthreads, do_wstacking):
-    
-    return _hdot_internal(uvw[0][0], freq, ms[0][0], wgt[0][0],
-                              freq_bin_idx, freq_bin_counts, nx, ny,
-                              cellx, celly, nu, nv, epsilon, nthreads,
-                              do_wstacking)
-
-def _hdot_internal(uvw, freq, ms, wgt, freq_bin_idx, freq_bin_counts,
-                   nx, ny, cellx, celly, nu, nv, epsilon, nthreads, do_wstacking):
-    freq_bin_idx -= freq_bin_idx.min()  # adjust for chunking
-    nband = freq_bin_idx.size
-    dirty = np.zeros((nband, nx, ny), dtype=freq.dtype)
-    for i in range(nband):
-        I = slice(freq_bin_idx[i], freq_bin_idx[i] + freq_bin_counts[i])
-        dirty[i] = ms2dirty(uvw=uvw, freq=freq[I], ms=ms[:, I, 0], wgt=wgt[:, I, 0], 
-                            npix_x=nx, npix_y=ny, pixsize_x=cellx, pixsize_y=celly,
-                            nu=nu, nv=nv, epsilon=epsilon, nthreads=nthreads, 
-                            do_wstacking=do_wstacking) + \
-                   ms2dirty(uvw=uvw, freq=freq[I], ms=ms[:, I, -1], wgt=wgt[:, I, -1], 
-                            npix_x=nx, npix_y=ny, pixsize_x=cellx, pixsize_y=celly,
-                            nu=nu, nv=nv, epsilon=epsilon, nthreads=nthreads, 
-                            do_wstacking=do_wstacking)
-    return dirty/2.0
-
-
-class wgridder(object):
-    """
-    Uses dask-ms to provide a chunked up interface to the measurement for the
-    wgridder in ducc. For simplicity it is assumed that the channel mapping
-    defines the chunking along the frequency axis where the channel mapping
-    is defined by the frequency resolution of the measurement and the number
-    of required imaging bands. By default it is assumed that all rows for a
-    specific imaging band fits into memory (i.e. row_chunks=-1). If this is
-    not the case row_chunks should be set explicitly (large chunks preferred).
-    The intended behaviour is for the chunks to be traversed in serial (to save
-    memory) and to leave the parallelisation to the gridder within each chunk.
-    
-    A list of measurement sets is supported but currently only a single phase
-    direction and frequency range with fixed number of channels are allowed. 
-    
-    Only even pixel sizes are currently supported. 
-
-    If a weighting scheme other than natural weighting is required they need
-    to be pre-computed and written to the measurement set in advance.
-
-    Only Stokes I maps are currently supported and these are computed by
-    gridding the first and last correlations separately and then combining
-    them. This doubles the amount of time spent gridding but it is the
-    correct way to do it (LB - TODO - should test this).
-
-    """
-    def __init__(self, ms, nx, ny, cell_size, cell_size_y=None, nband=None, field=0,
-                 ddid=0, precision=1e-7, nthreads=0, do_wstacking=1, row_chunks=-1,
-                 data_column='DATA', weight_column='WEIGHT_SPECTRUM',
-                 model_column="MODEL_DATA", flag_column=None, psf_oversize=2.0,
-                 weight_norm='backward', padding=2.0, out_stokes='I'):
-        """
-        Parameters
-        ----------
-        ms : list
-            List of measurement sets
-        nx : int
-            Number of pixels along l direction.
-        ny : int
-            Number of pixels along m dimension.
-        cell_size : float
-            Cell size in arcseconds.  
-        cell_size_y : float, optional
-            Cell size in m direction in arcseconds. 
-            If None the pixels will be square.  
-        nband : int, optional
-            Number of imaging bands. 
-            If None then the image cube will have the same
-            frequency resolution as the measurement set.
-        field : int, optional
-            Which field to image. Default is field 0 for all
-            measurement sets.
-            All fields currently need to have the same phase direction.
-        ddid : int, optional
-            Which DDID to image. Defaults to 0 for all
-            measurement sets.
-        precision : float, optional
-            Gridder error tolerance. Defaults to 1e-7.
-        nthreads : int, optional.
-            The number of threads to give the gridder.
-            Default of 0 means use all threads.
-        do_wstacking : bool, optional
-            Whether to perform w-stacking or not. 
-            Defaults to True.
-        row_chunks : int, optional
-            Row chunking per imaging band.
-            Default assumes all rows for an
-            imaging band fits into memory.
-        data_column : string, optional
-            Which measurement set column to image. 
-            Defaults to 'DATA'.
-        weight_column : string, optional
-            Defaults to 'WEIGHT_SPECTRUM' but will use 'WEIGHT' if it
-            does not exist. Custom columns need to have the same
-            dimensions as data. 
-        model_column : string, optional
-            Which column to write model visibilities to. 
-            The default column is 'MODEL_DATA' but the user needs to provide
-            a switch in the self.dot method to activate this. 
-            This is to avoid accidentally over writing an existing column.
-        flag_column : string, optional
-            The default of None will take the union of FLAG_ROW and FLAG.
-            For different behaviour point to a specific row. 
-            Note that flags are implemented by zeroing the corresponding
-            weights.
-        weight_norm : string, 'backward' or 'ortho'
-            This determines how the weights are applied. 
-            The default of 'backward' only applies the weights in the
-            backward (i.e. hdot) transform. This is the classical way
-            to apply the weights but it results in the forward and backward
-            transforms not being self adjoint. The 'ortho' mode applies the
-            square root of the weights in both directions and therefore
-            results in consistent operators. Note that in this mode the
-            data is also whitened and care has to be taken when
-            writing model visibilities to the measurement set.  
-        padding : float, optional
-            How much to pad by during gridding.
-            Defaults to 2
-        out_stokes : string, optional
-            Only Stokes I maps currently supported.
-
-        Methods
-        -------
-
-        dot : 
-            The forward transform from image to visibilities (wraps dirty2ms).
-
-        hdot : 
-            The backward transform from visibilities to image (wraps ms2dirty).
-            Can be used to make the dirty image
-            
-        make_residual : 
-            Compute residual image in place given a model image.
-
-        make_psf :
-            Computes an over-sized psf
-
-        TODO:
-            * more flexible freq mapping
-            * compute the chi-square
-            * different phase centers and beam interpolation
-
-
-        """
-        if out_stokes != 'I':
-            raise NotImplementedError("Only Stokes I maps currently supported")
-        if precision > 1e-6:
-            self.real_type = np.float32
-            self.complex_type = np.complex64
-        else:
-            self.real_type = np.float64
-            self.complex_type=np.complex128
-
-        # image size
-        self.nx = nx
-        assert self.nx % 2 == 0
-        self.ny = ny
-        assert self.ny % 2 == 0
-        nu = int(padding * self.nx)
-        self.nu = next_fast_len(nu)
-        nv = int(padding * self.ny)
-        self.nv = next_fast_len(nv)
-        # psf size
-        nx_psf = int(self.nx*psf_oversize)
-        if nx_psf%2:
-            self.nx_psf = nx_psf+1
-        else:
-            self.nx_psf = nx_psf
-        nu_psf = int(padding*self.nx_psf)
-        self.nu_psf = next_fast_len(nu_psf)
-        ny_psf = int(self.ny*psf_oversize)
-        if ny_psf%2:
-            self.ny_psf = ny_psf+1
-        else:
-            self.ny_psf = ny_psf
-        nv_psf = int(padding*self.ny_psf)
-        self.nv_psf = next_fast_len(nv_psf)
-        self.cell = cell_size * np.pi/60/60/180
-        if cell_size_y is not None:
-            self.celly = cell_size_y * np.pi/60/60/180
-        else:
-            self.celly = self.cell
-        self.field = field
-        self.ddid = ddid
-        self.precision = precision
-        self.nthreads = nthreads
-        self.do_wstacking = do_wstacking
-        if weight_norm != 'backward' and weight_norm != 'ortho':
-            raise ValueError("Unknown weight_norm supplied")
-        else:
-            self.weight_norm = weight_norm
-
-        if isinstance(ms, list):
-            self.ms = ms
-        else:
-            self.ms = [ms]
-        
-        # first we check that phase centres, DDID freqs and pols are the same
-        freq = None
-        ptype = None
-        radec = None
-        for ims in self.ms:
-            xds = xds_from_ms(ims, chunks={'row':-1})
-
-            if not weight_column in xds[0].data_vars:
-                print('%s not in ms. Using WEIGHT instead' % weight_column)
-                self.weight_column = 'WEIGHT'
-            else:
-                self.weight_column = weight_column
-
-            # Get subtable data
-            ddids = xds_from_table(ims + "::DATA_DESCRIPTION", group_cols="__row__")[self.ddid]
-            fields = xds_from_table(ims + "::FIELD", group_cols="__row__")[self.field]
-            spws = xds_from_table(ims + "::SPECTRAL_WINDOW", group_cols="__row__")
-            pols = xds_from_table(ims + "::POLARIZATION", group_cols="__row__")
-
-            ddid = dask.compute(ddids)[0]
-            field = dask.compute(fields)[0]
-            spws = dask.compute(spws)[0]
-            pols = dask.compute(pols)[0]
-
-            for ds in xds:
-                if ds.FIELD_ID != self.field and ds.DATA_DESC_ID != self.ddid:
-                    continue
-
-                # check correct phase direction
-                if radec is None:
-                    radec = field.PHASE_DIR.data.squeeze()
-                else: 
-                    assert_array_equal(radec, field.PHASE_DIR.data.squeeze())
-
-                # same polarisation type
-                pol = pols[ddid.POLARIZATION_ID.values[0]]
-                corr_type_set = set(pol.CORR_TYPE.data.squeeze())
-                if corr_type_set.issubset(set([9, 10, 11, 12])):
-                    pol_type = 'linear'
-                elif corr_type_set.issubset(set([5, 6, 7, 8])):
-                    pol_type = 'circular'
-                else:
-                    raise ValueError("Cannot determine polarisation type "
-                                    "from correlations %s. Constructing "
-                                    "a feed rotation matrix will not be "
-                                    "possible." % (corr_type_set,))
-
-                if ptype is None:
-                    ptype = pol_type
-                else:
-                    assert ptype == pol_type
-
-                # same frequencies            
-                spw = spws[ddid.SPECTRAL_WINDOW_ID.values[0]]
-                if freq is None:
-                    freq = spw.CHAN_FREQ.data.squeeze()
-                else:
-                    assert_array_equal(freq, spw.CHAN_FREQ.data.squeeze())
-
-        self.freq = freq
-        self.ptype = ptype
-
-        # compute freq mapping and channel chunks
-        self.nchan = freq.size
-        if nband is None:
-            self.nband = self.nchan
-        else:
-            self.nband = nband
-        step = self.nchan//self.nband
-        freq_mapping = np.arange(0, self.nchan, step)
-        self.freq_mapping = np.append(freq_mapping, self.nchan)
-        self.freq_out = np.zeros(self.nband)
-        for i in range(self.nband):
-            Ilow = self.freq_mapping[i]
-            Ihigh = self.freq_mapping[i+1]
-            self.freq_out[i] = np.mean(self.freq[Ilow:Ihigh])
-        self.chan_chunks = list(self.freq_mapping[1::] - self.freq_mapping[0:-1])
-        self.freq_da = da.from_array(self.freq, chunks=self.chan_chunks)
-        self.freq_bin_idx = da.from_array(self.freq_mapping[0:-1], chunks=1)
-        self.freq_bin_counts = da.from_array(self.freq_mapping[1::] - self.freq_mapping[0:-1], chunks=1)
-
-        ######################################################################################################################
-        # meta info for xds_from_table
-        self.data_column = data_column
-        self.model_column = model_column
-        self.ms = ms
-        self.row_chunks = row_chunks
-        # self.schema = {
-        #     self.data_column: {'dims': ('chan',)},
-        #     self.weight_column: {'dims': ('chan', )},
-        #     "UVW": {'dims': ('uvw',)},
-        # }
-
-    def dot(self, x, column=None):
-        """
-        Implements forward transform i.e.
-
-        V = Rx
-
-        where R is the interferometric response.
-        It is assumed that the result does not fit in memory. 
-        For this reason it is always written to column in the MS.  
-        """
-        if column is None:
-            column = self.model_column
-        model = da.from_array(x, chunks=(1, self.nx, self.ny))
-        writes = []
-        for ims in self.ms:
-            xds = xds_from_ms(ims, columns=('UVW'), chunks={"row":self.row_chunks, "chan": self.chan_chunks})
-            out_data = []
-            for ds in xds:
-                # if ds.FIELD_ID != self.field and ds.DATA_DESC_ID != self.ddid:
-                #     continue
-                uvw = ds.UVW.data
-                ncorr = 4 #getattr(ds, column).shape[-1]
-
-                vis = da.blockwise(_dot_wrapper, ('row', 'chan', 'corr'),
-                                   uvw, ('row', 'three'), 
-                                   self.freq_da, ('chan',),
-                                   model, ('chan', 'nx', 'ny'),
-                                   self.freq_bin_idx, ('chan',),
-                                   self.freq_bin_counts, ('chan',),
-                                   self.cell, None, 
-                                   self.celly, None,
-                                   self.nu, None, 
-                                   self.nv, None,
-                                   self.precision, None,
-                                   self.nthreads, None,
-                                   self.do_wstacking, None,
-                                   ncorr, None,
-                                   self.complex_type, None,
-                                   adjust_chunks={'chan': self.freq_da.chunks[0]},
-                                   new_axes={"corr": ncorr},
-                                   dtype=self.complex_type, 
-                                   align_arrays=False)
-                
-                data_vars = {column:(('row', 'chan', 'corr'), vis)}
-                out_ds = Dataset(data_vars)
-                out_data.append(out_ds)
-                # print(vis.max(), vis.min())
-                # ds.assign(**{column: (("row", "chan", "corr"), vis)})
-            writes.append(xds_to_table(out_data, ims, columns=[column]))
-        dask.compute(writes, scheduler='single-threaded')
-
-    def hdot(self, data_column=None, weight_column=None):
-        """
-        Implements the backward transform
-
-        I^D = R.H V
-
-        It is assumed that the result always fits in memory.
-        If model_data is not passed in then it produces the
-        dirty image.
-        """
-        if data_column is None:
-            data_column = self.data_column
-        if weight_column is None:
-            weight_column = self.weight_column
-        dirty = da.zeros((self.nband, self.nx, self.ny), chunks=(1, self.nx, self.ny), dtype=self.real_type)
-        for ims in self.ms:
-            xds = xds_from_ms(ims, columns=(data_column, weight_column, 'UVW'), chunks={"row":self.row_chunks, "chan": self.chan_chunks})
-            for ds in xds:
-                # if ds.FIELD_ID.data != self.field and ds.DATA_DESC_ID.data != self.ddid:
-                #     continue
-                data = getattr(ds, data_column).data
-                weights = getattr(ds, weight_column).data
-                if weights.shape != data.shape:
-                    weights = da.broadcast_to(weights[:, None, :], data.shape)
-                    weights = da.rechunk(weights, data.chunks)
-                uvw = ds.UVW.data
-
-                dirty += da.blockwise(_hdot_wrapper, ('chan', 'nx', 'ny'),
-                                      uvw, ('row', 'three'), 
-                                      self.freq_da, ('chan',),
-                                      data, ('row', 'chan', 'corr'),
-                                      weights, ('row', 'chan', 'corr'),
-                                      self.freq_bin_idx, ('chan',),
-                                      self.freq_bin_counts, ('chan',),
-                                      self.nx, None,
-                                      self.ny, None, 
-                                      self.cell, None, 
-                                      self.celly, None,
-                                      self.nu, None, 
-                                      self.nv, None,
-                                      self.precision, None,
-                                      self.nthreads, None,
-                                      self.do_wstacking, None,
-                                      adjust_chunks={'chan': self.freq_bin_idx.chunks[0]},
-                                      new_axes={"nx": self.nx, "ny": self.ny},
-                                      dtype=self.real_type, 
-                                      align_arrays=False)
-
-        return dirty.compute(scheduler='single-threaded')
-
-
-    def make_residual(self, x, v_dof=None):
-        print("Making residual")
-        residual = np.zeros(x.shape, dtype=x.dtype)
-        for ims in self.ms:
-            xds = xds_from_ms(ims, chunks={"row":-1, "chan": self.chan_chunks}, table_schema=self.schema)
-            for ds in xds:
-                # if ds.FIELD_ID.data != self.field and ds.DATA_DESC_ID.data != self.ddid:
-                #     print('In here')
-                #     continue
-                data = getattr(ds, self.data_column).data
-                weights = getattr(ds, self.weight_column).data
-                uvw = ds.UVW.data.astype(self.real_type)
-
-                for i in range(self.nband):
-                    Ilow = self.freq_mapping[i]
-                    Ihigh = self.freq_mapping[i+1]
-                    weighti = weights.blocks[:, i].compute().astype(self.real_type)
-                    datai = data.blocks[:, i].compute().astype(self.complex_type)
-
-                    # TODO - load and apply interpolated fits beam patterns for field
-
-                    # get residual vis
-                    if weighti.any():
-                        residual_vis = weighti * datai - ng.dirty2ms(uvw=uvw, freq=self.freq[Ilow:Ihigh], dirty=x[i], wgt=weighti,
-                                                                    pixsize_x=self.cell, pixsize_y=self.cell, epsilon=self.precision,
-                                                                    nthreads=self.nthreads, do_wstacking=self.do_wstacking, verbosity=0)
-
-                        # make residual image
-                        residual[i] += ng.ms2dirty(uvw=uvw, freq=self.freq[Ilow:Ihigh], ms=residual_vis, wgt=weighti,
-                                                npix_x=self.nx, npix_y=self.ny, pixsize_x=self.cell, pixsize_y=self.cell,
-                                                epsilon=self.precision, nthreads=self.nthreads, do_wstacking=self.do_wstacking, verbosity=0)
-        return residual
-
-    
-
-    def make_psf(self, weight_column=None):
-        """
-        Computes over-sized psf
-        """
-        if weight_column is None:
-            weight_column = self.weight_column
-        psf = da.zeros((self.nband, self.nx_psf, self.ny_psf), chunks=(1, self.nx_psf, self.ny_psf), dtype=self.real_type)
-        for ims in self.ms:
-            xds = xds_from_ms(ims, columns=(weight_column, 'UVW'), chunks={"row":self.row_chunks, "chan": self.chan_chunks})
-            for ds in xds:
-                # print(ds.FIELD_ID.data, self.field)
-                # print(ds.DATA_DESC_ID.data, self.ddid)
-                # if ds.FIELD_ID.data != self.field and ds.DATA_DESC_ID.data != self.ddid:
-                #     print('In here')
-                #     continue
-                weights = getattr(ds, weight_column).data
-                if len(weights.shape) < 3:
-                    nrow, ncorr = weights.shape
-                    weights = da.broadcast_to(weights[:, None, :], (nrow, self.nchan, ncorr))
-                    weights = da.rechunk(weights, (self.row_chunks, self.chan_chunks, ncorr))
-                uvw = ds.UVW.data
-
-                wgt = da.sqrt(weights)
-                psf += da.blockwise(_hdot_wrapper, ('chan', 'nx', 'ny'),
-                                    uvw, ('row', 'three'), 
-                                    self.freq_da, ('chan',),
-                                    wgt.astype(self.complex_type), ('row', 'chan', 'corr'),
-                                    wgt, ('row', 'chan', 'corr'),
-                                    self.freq_bin_idx, ('chan',),
-                                    self.freq_bin_counts, ('chan',),
-                                    self.nx_psf, None,
-                                    self.ny_psf, None, 
-                                    self.cell, None, 
-                                    self.celly, None,
-                                    self.nu_psf, None, 
-                                    self.nv_psf, None,
-                                    self.precision, None,
-                                    self.nthreads, None,
-                                    self.do_wstacking, None,
-                                    adjust_chunks={'chan': self.freq_bin_idx.chunks[0]},
-                                    new_axes={"nx": self.nx_psf, "ny": self.ny_psf},
-                                    dtype=self.real_type, 
-                                    align_arrays=False)
-
-        return psf.compute(scheduler='single-threaded')
 
 
 class GridderConfigWrapper(object):
